feat: added image ID verification

+++ backend/routes/profile.go
@@ -21,6 +21,7 @@ func (con controller) Profile(c *gin.Context) {
 type ProfileUpdateData struct {
 	DisplayName string `json:"display_name,omitempty"`
 	ImageData   string `json:"image_data,omitempty"`
+	IDCardData  string `json:"id_card_data,omitempty"`
 }
 
 // ProfileUpdate updates the user's profile.
@@ -48,7 +49,7 @@ func (con controller) ProfileUpdate(c *gin.Context) {
 		}
 	}
 
-	if userUpdateData.ImageData == "" {
+	if userUpdateData.ImageData == "" && userUpdateData.IDCardData == "" {
 		if displayNameUpdate {
 			c.JSON(http.StatusOK, gin.H{"user": user.Sanitize()})
 			return
@@ -57,20 +58,40 @@ func (con controller) ProfileUpdate(c *gin.Context) {
 		c.JSON(http.StatusBadRequest, gin.H{"error": "no data to update"})
 	}
 
-	url, err := con.uploader.UploadFile(userUpdateData.ImageData)
-	if err != nil {
-		c.JSON(http.StatusInternalServerError, gin.H{"error": "error uploading"})
-		return
+	if userUpdateData.ImageData != "" {
+		url, err := con.uploader.UploadFile(userUpdateData.ImageData)
+		if err != nil {
+			c.JSON(http.StatusInternalServerError, gin.H{"error": "error uploading"})
+			return
+		}
+
+		if err = con.uploader.DeleteFile(user.Profile.ImageURL); err != nil {
+			log.Println("error deleting old image:", err)
+		}
+
+		user.Profile.ImageURL = url
+		if res := con.db.Model(user.Profile).Where("user_id = ?", user.ID).Updates(user.Profile); res.Error != nil {
+			c.JSON(http.StatusInternalServerError, gin.H{"error": "error saving"})
+			return
+		}
 	}
 
-	if err = con.uploader.DeleteFile(user.Profile.ImageURL); err != nil {
-		log.Println("error deleting old image:", err)
-	}
+	if userUpdateData.IDCardData != "" {
+		url, err := con.uploader.UploadFile(userUpdateData.IDCardData)
+		if err != nil {
+			c.JSON(http.StatusInternalServerError, gin.H{"error": "error uploading"})
+			return
+		}
 
-	user.Profile.ImageURL = url
-	if res := con.db.Model(user.Profile).Where("user_id = ?", user.ID).Updates(user.Profile); res.Error != nil {
-		c.JSON(http.StatusInternalServerError, gin.H{"error": "error saving"})
-		return
+		if err = con.uploader.DeleteFile(user.Profile.UnsecureIDCardURL); err != nil {
+			log.Println("error deleting old image:", err)
+		}
+
+		user.Profile.UnsecureIDCardURL = url
+		if res := con.db.Model(user.Profile).Where("user_id = ?", user.ID).Updates(user.Profile); res.Error != nil {
+			c.JSON(http.StatusInternalServerError, gin.H{"error": "error saving"})
+			return
+		}
 	}
 
 	c.JSON(http.StatusOK, gin.H{"user": user.Sanitize()})
